requirements.txt
# AWFM Backend Requirements
# Django 6.0 + Django REST Framework

# Core
Django==6.0
psycopg2-binary==2.9.10  # PostgreSQL adapter

# Django REST Framework
djangorestframework==3.15.2
django-filter==24.3  # Filtering support
drf-spectacular==0.28.0  # OpenAPI/Swagger documentation

# CORS
django-cors-headers==4.6.0

# Authentication & Security
PyJWT==2.10.1  # JWT tokens for auth
djangorestframework-simplejwt==5.3.1  # JWT authentication for DRF
bcrypt==4.1.2  # Bcrypt password hashing

# Email
sendgrid==6.11.0  # SendGrid email service

# Development
python-dotenv==1.0.1  # Environment variables

# Production
gunicorn==23.0.0  # WSGI server
whitenoise==6.8.2  # Static file serving
daphne==4.1.2  # ASGI server for WebSockets

# Media Storage (Cloudinary)
cloudinary==1.41.0  # Cloudinary SDK
django-cloudinary-storage==0.3.0  # Django integration for Cloudinary
Pillow==11.0.0  # Image processing (required by django-cloudinary-storage)

# Redis & WebSockets (Django Channels)
redis==5.2.1  # Redis client
channels==4.2.0  # WebSocket support
channels-redis==4.2.1  # Redis channel layer for Channels

# Background tasks (optional, uncomment when needed)
# celery==5.4.0  # Background tasks

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0
factory-boy==3.3.1  # Test factories for model instances
faker==33.1.0  # Generate fake data for tests
